Refactor ContestHeader into it's own component
