+++ packages/app/pages/contest/[id].tsx
@@ -1,38 +1,15 @@
 import { CheckCircleFilled, CloseCircleFilled } from "@ant-design/icons";
-import { Card, Progress, Tabs } from "antd";
+import { Card, Tabs } from "antd";
 import Table, { ColumnsType } from "antd/lib/table";
-import Text from "antd/lib/typography/Text";
-import Title from "antd/lib/typography/Title";
-import dayjs from "dayjs";
-import duration from "dayjs/plugin/duration";
 import { GetServerSideProps, InferGetServerSidePropsType } from "next";
-import { HTMLAttributes, useState } from "react";
+import { HTMLAttributes } from "react";
 import classNames from "classnames";
-import { useInterval } from "../../hooks/useInterval";
-import { useTwoPassRendering } from "../../hooks/useTwoPassRendering";
 import { Hyperlink } from "../../components/Hyperlink";
-
-dayjs.extend(duration);
+import { ContestHeader } from "../../components/ContestHeader";
 
 type ProblemStatus = "not submitted" | "accepted" | "rejected";
 
 const { TabPane } = Tabs;
-
-interface ProblemOverview {
-  id: string;
-  origin: string;
-  title: string;
-  status: ProblemStatus;
-  solvedCount: number;
-  attemptedCount: number;
-}
-
-interface Contest {
-  title: string;
-  startDate: string;
-  endDate: string;
-  problems: ProblemOverview[];
-}
 
 interface ContestResponse {
   data: Contest;
@@ -154,91 +131,6 @@ export const getServerSideProps: GetServerSideProps<ContestResponse> = async () 
     },
   },
 });
-
-type ContestHeaderProps = Pick<Contest, "startDate" | "endDate" | "title">;
-
-const ContestHeader = ({
-  startDate: rawStartDate,
-  endDate: rawEndDate,
-  title,
-}: ContestHeaderProps) => {
-  const isClient = useTwoPassRendering();
-
-  const startDate = dayjs(rawStartDate);
-  const endDate = dayjs(rawEndDate);
-  const formattedStartDate = startDate.format("YYYY-MM-DD HH:mm Z UTC");
-  const formattedEndDate = endDate.format("YYYY-MM-DD HH:mm Z UTC");
-
-  const [now, setNow] = useState(dayjs());
-
-  const getTimeDifferenceFrom = (date: dayjs.Dayjs) => {
-    const timeDifference = dayjs.duration(
-      date.isBefore(now) ? now.diff(date) : date.diff(now),
-    );
-    const days = timeDifference.days();
-
-    if (days > 1) {
-      return `${days} dias`;
-    }
-
-    if (days === 1) {
-      return `${Math.floor(timeDifference.asHours())} horas`;
-    }
-
-    return timeDifference.format("HH:mm:ss");
-  };
-
-  const contestStatus = (() => {
-    if (now.isBefore(startDate)) return "scheduled";
-    if (now.isBefore(endDate)) return "running";
-    return "ended";
-  })();
-
-  const relativeTimeInfo = (() => {
-    if (contestStatus === "scheduled") {
-      return `Iniciará em ${getTimeDifferenceFrom(startDate)}`;
-    }
-    if (contestStatus === "running") {
-      return `Tempo restante: ${getTimeDifferenceFrom(endDate)}`;
-    }
-    return <Text type="success">Acabou</Text>;
-  })();
-
-  const percentage = (() => {
-    if (contestStatus === "scheduled") return 0;
-    if (contestStatus === "ended") return 100;
-    return +(100 * (now.diff(startDate) / endDate.diff(startDate))).toFixed(2);
-  })();
-
-  useInterval(() => setNow(dayjs()), contestStatus !== "ended" ? 1000 : null);
-
-  return (
-    <Card
-      className="card"
-      bodyStyle={{ paddingTop: 16 }}
-      title={<Title level={1}>{title}</Title>}
-    >
-      <Progress
-        percent={isClient ? percentage : 0}
-        status={contestStatus === "running" ? "active" : undefined}
-        showInfo={false}
-      />
-      <div className="flex flex-col md:flex-row md:justify-between md:items-center">
-        <div>
-          <Text strong>Início: </Text>
-          {formattedStartDate}
-        </div>
-        <div className="order-first md:order-none">
-          <Title level={5}>{isClient ? relativeTimeInfo : "..."}</Title>
-        </div>
-        <div>
-          <Text strong>Término: </Text>
-          {formattedEndDate}
-        </div>
-      </div>
-    </Card>
-  );
-};
 
 interface ContestProblemsProps {
   problems: ProblemOverview[];
