+++ packages/app/typings/global.d.ts
@@ -1,0 +1,15 @@
+declare interface ProblemOverview {
+  id: string;
+  origin: string;
+  title: string;
+  status: ProblemStatus;
+  solvedCount: number;
+  attemptedCount: number;
+}
+
+declare interface Contest {
+  title: string;
+  startDate: string;
+  endDate: string;
+  problems: ProblemOverview[];
+}
